Reduce bluetooth flapping on RasPi rigs.

+++ bin/oref0-bluetoothup.sh
@@ -5,9 +5,14 @@ if ! ( ps -fC bluetoothd ) ; then
    sudo /usr/local/bin/bluetoothd &
 fi
 
-if ! ( hciconfig -a | grep -q "PSCAN" ) ; then
-   sudo killall bluetoothd
-   sudo /usr/local/bin/bluetoothd &
+#Raspberry Pi doesn't keep PSCAN up the way Edison does. Check for ARM CPU (vs x86 on Edison) before executing this bloc
+if ! ( uname -m | grep arm > /dev/null)
+then
+   if ! ( hciconfig -a | grep -q "PSCAN" ) ; then
+      echo "On Edison and no PSCAN - restarting bluetoothd"
+      sudo killall bluetoothd
+      sudo /usr/local/bin/bluetoothd &
+   fi
 fi
 
 if ( hciconfig -a | grep -q "DOWN" ) ; then
